+++ streams/events/src/errors.rs
@@ -1,0 +1,5 @@
+error_chain! {
+    links {
+        StoreError(::replicante_data_store::Error, ::replicante_data_store::ErrorKind);
+    }
+}